Thunks y ficheros relacionados con el Login refactorizados para usar la API. Ahora las rutas estan guardadas de forma privada en el fichero .env

+++ src/contact/features/thunks/contactDeleteByIdThunk.ts
@@ -1,13 +1,15 @@
 
 import { createAsyncThunk } from "@reduxjs/toolkit"
-import { apiUrl, apiEndPointContacts, apiToken } from "../../../common/globalParameters/routes.ts"
 
 
 export const ContactDeleteByIdThunk = createAsyncThunk
     ("contact/deleteById", async (contactId: string) => {
 
+        const apiToken = localStorage.getItem('token')
+        if (!apiToken) return '0'
+
         try {
-            const request = await fetch(`${apiUrl}/${apiEndPointContacts}/${contactId}`, {
+            const request = await fetch(`${import.meta.env.VITE_API_URL}/${import.meta.env.VITE_API_ENDPOINT_CONTACTS}/${contactId}`, {
                 method: "DELETE",
                 headers: {
                     "Content-Type": "application/json",
